ACC-138: define and apply docker exo funstions

+++ _functions.sh
@@ -32,6 +32,7 @@ source "${SCRIPT_DIR}/_functions_chat.sh"
 source "${SCRIPT_DIR}/_functions_onlyoffice.sh"
 source "${SCRIPT_DIR}/_functions_ldap.sh"
 source "${SCRIPT_DIR}/_functions_cmis.sh"
+source "${SCRIPT_DIR}/_functions_exo_docker.sh"
 
 # #################################################################################
 #
@@ -279,7 +280,9 @@ initialize_product_settings() {
       configurable_env_var "USER_DIRECTORY_ADMIN_DN" "cn=admin,dc=exoplatform,dc=com"
       configurable_env_var "USER_DIRECTORY_ADMIN_PASSWORD" "exo"
       # Deploy exo in docker or not
-      configurable_env_var "DEPLOY_EXO_DOCKER" false      
+      configurable_env_var "DEPLOY_EXO_DOCKER" false  
+      configurable_env_var "DEPLOYMENT_EXO_DOCKER_IMAGE" ""
+      configurable_env_var "DEPLOYMENT_EXO_DOCKER_IMAGE_VERSION" ""   
 
       if [[ "$DEPLOYMENT_ADDONS" =~ "exo-onlyoffice" ]]; then
         env_var "DEPLOYMENT_ONLYOFFICE_DOCUMENTSERVER_ENABLED" true
@@ -527,6 +530,15 @@ initialize_product_settings() {
             env_var PLF_BRANCH "${PRODUCT_BRANCH} Demo"
           else
             env_var PLF_BRANCH "${PRODUCT_BRANCH}"
+          fi
+        ;;
+        plfdocker)
+          env_var PRODUCT_DESCRIPTION "Platform EE in docker"
+          env_var DEPLOY_EXO_DOCKER "true"
+          env_var DEPLOYMENT_EXO_DOCKER_IMAGE_VERSION "${PRODUCT_VERSION}"          
+          if [ -z "${DEPLOYMENT_EXO_DOCKER_IMAGE}" ] || [ -z "${DEPLOYMENT_EXO_DOCKER_IMAGE_VERSION}" ] ; then
+           echo_error "DEPLOYMENT_EXO_DOCKER_IMAGE (or/and) DEPLOYMENT_EXO_DOCKER_IMAGE_VERSION have to be specified with docker deployment"            
+           exit 1
           fi
         ;;
         plfenteap)
@@ -777,7 +789,7 @@ initialize_product_settings() {
           fi
 
           # For configuration differences between tomcat and jboss
-          if [[ "${PRODUCT_NAME}" =~ ^(plfcom|plfent|plfentrial|plfsales)$ ]]; then
+          if [[ "${PRODUCT_NAME}" =~ ^(plfcom|plfent|plfentrial|plfsales|plfdocker)$ ]]; then
             if [[ "${PRODUCT_VERSION}" =~ ^(5.0|5.1|5.2|5.3|6.0) ]]; then
               env_var "DEPLOYMENT_APPSRV_VERSION" "8.5"
             else 
@@ -853,11 +865,12 @@ initialize_product_settings() {
 
    do_get_plf_settings
    do_get_cmis_settings
-   do_get_onlyoffice_settings
+   do_get_onlyoffice_settings   
    do_get_ldap_settings
    do_get_database_settings
    do_get_es_settings
    do_get_chat_settings
+   do_get_exo_docker_settings
 }
 
 #
@@ -915,6 +928,11 @@ do_download_dataset() {
 }
 
 do_restore_dataset(){
+
+  if ${DEPLOY_EXO_DOCKER}; then
+    echo_error "restore dataset not yet implemented for docker deployment"
+    exit 1
+  fi  
   # System dependent settings
   if ${LINUX}; then
     env_var "TAR_BZIP2_COMPRESS_PRG" "--use-compress-prog=pbzip2"
@@ -963,12 +981,16 @@ do_init_empty_data(){
 # Drops all data used by the instance.
 #
 do_drop_data() {
-  echo_info "Drops instance indexes ..."
-  rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}/jcr/index/
-  echo_info "Done."
-  echo_info "Drops instance values ..."
-  rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}/jcr/values/
-  echo_info "Done."
+  if ${DEPLOY_EXO_DOCKER}; then
+   do_drop_exo_docker_data
+  else
+   echo_info "Drops instance indexes ..."
+   rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}/jcr/index/
+   echo_info "Done."
+   echo_info "Drops instance values ..."
+   rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}/jcr/values/
+   echo_info "Done."
+  fi
   echo_info "Drops instance codec folder ..."
   rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR}/
   echo_info "Done."
@@ -1053,15 +1075,19 @@ do_unpack_server() {
 # Creates all data directories used by the instance.
 #
 do_create_data() {
-  echo_info "Creates instance indexes directory ..."
-  mkdir -p ${DEPLOYMENT_DIR}/gatein/data/jcr/index/
-  echo_info "Done."
-  echo_info "Creates instance values directory ..."
-  mkdir -p ${DEPLOYMENT_DIR}/gatein/data/jcr/values/
-  echo_info "Done."
-  echo_info "Creates instance codec directory ..."
-  mkdir -p ${DEPLOYMENT_DIR}${DEPLOYMENT_CODEC_DIR}
-  echo_info "Done."
+  if ${DEPLOY_EXO_DOCKER}; then
+   do_create_exo_docker
+  else
+   echo_info "Creates instance indexes directory ..."
+   mkdir -p ${DEPLOYMENT_DIR}/gatein/data/jcr/index/
+   echo_info "Done."
+   echo_info "Creates instance values directory ..."
+   mkdir -p ${DEPLOYMENT_DIR}/gatein/data/jcr/values/
+   echo_info "Done."
+  fi
+   echo_info "Creates instance codec directory ..."
+   mkdir -p ${DEPLOYMENT_DIR}${DEPLOYMENT_CODEC_DIR}
+   echo_info "Done."
 }
 
 do_configure_apache() {
@@ -1370,54 +1396,33 @@ do_docker_server_deploy(){
     do_stop
   fi
   if [ "${DEPLOYMENT_MODE}" == "KEEP_DATA" ]; then
-    echo_info "Archiving existing data ${INSTANCE_DESCRIPTION} ..."
-    _tmpdir=`mktemp -d -t archive-data.XXXXXXXXXX` || exit 1
-    echo_info "Using temporary directory ${_tmpdir}"
     if [ ! -e "${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}" ]; then
       echo_warn "This instance wasn't deployed before. Nothing to keep."
-      mkdir -p ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR})
-      mkdir -p ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR})
+      do_create_exo_docker
       do_create_database
       do_create_chat_database
       do_create_es
       do_create_onlyoffice
       do_create_cmis
     else
-      # Use a subshell to not expose settings loaded from the deployment descriptor
-      (
-      # The server have been already deployed.
-      # We load its settings from the configuration
-      do_load_deployment_descriptor
-      if [ -d "${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}" ]; then
-        mv ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR} ${_tmpdir}
-        mv ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR} ${_tmpdir}
-      else
-        mkdir -p ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR})
-        mkdir -p ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR})
-        do_create_database
-        do_create_chat_database
-        do_create_es
-        do_create_onlyoffice
-        do_create_cmis
-      fi
-      )
+     do_load_deployment_descriptor
     fi
-    echo_info "Done."
+  else  
+    # Use a subshell to not expose settings loaded from the deployment descriptor
+    (
+    # The server have been already deployed.
+    # We load its settings from the configuration     
+    do_create_exo_docker
+    do_create_database
+    do_create_chat_database
+    do_create_es
+    do_create_onlyoffice
+    do_create_cmis
+    
+    )
   fi
-
-  set -e
-  DEPLOYMENT_PID_FILE=${SRV_DIR}/${INSTANCE_KEY}.pid
-  mkdir -p ${SRV_DIR}
-  echo_info "Deleting existing server ..."
-  rm -rf ${SRV_DIR}/${INSTANCE_KEY}
-  echo_info "Done"
-  cp -rf ${TMP_DIR}/${INSTANCE_KEY} ${SRV_DIR}/${INSTANCE_KEY}
-  rm -rf ${TMP_DIR}/${INSTANCE_KEY}
-
-  # We search the server directory
-  pushd `find ${SRV_DIR}/${INSTANCE_KEY} -maxdepth 4 -mindepth 1 -name bin -type d`/.. > /dev/null
-  DEPLOYMENT_DIR=`pwd -P`
-  popd > /dev/null
+  echo_info "Done."
+  
 
   # Initialize database before configuratation
   # before with docker the datase must be started before to retreive the port number
@@ -1426,13 +1431,7 @@ do_docker_server_deploy(){
       do_init_empty_data
     ;;
     KEEP_DATA)
-      echo_info "Restoring previous data ${INSTANCE_DESCRIPTION} ..."
-      rm -rf ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}
-      mkdir -p $(dirname ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR})
-      mv ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}) ${DEPLOYMENT_DIR}/${DEPLOYMENT_DATA_DIR}
-      mv ${_tmpdir}/$(basename ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR}) ${DEPLOYMENT_DIR}/${DEPLOYMENT_CODEC_DIR}
-      rm -rf ${_tmpdir}
-      echo_info "Done."
+      echo_info "previous data ${INSTANCE_DESCRIPTION} are already kept ..."      
     ;;
     RESTORE_DATASET)
       do_restore_dataset
@@ -1443,8 +1442,6 @@ do_docker_server_deploy(){
       exit 1
     ;;
   esac
-  do_get_tomcat_settings
-  do_configure_tomcat_server 
 
 }
 
@@ -1452,6 +1449,14 @@ do_docker_server_deploy(){
 # Function that starts the app server
 #
 do_start() {
+  if ${DEPLOY_EXO_DOCKER}; then
+   do_docker_start
+  else
+   do_default_start
+  fi
+
+}
+do_default_start() {
   # Use a subshell to not expose settings loaded from the deployment descriptor
   (
   # The server is supposed to be already deployed.
@@ -1583,10 +1588,37 @@ do_start() {
   )
 }
 
+do_docker_start(){
+  (
+  do_start_onlyoffice
+  do_start_ldap
+  do_start_cmis
+  do_start_database
+  do_start_es
+  do_start_chat_server
+
+  # We need this variable for the setenv
+  export DEPLOYMENT_CHAT_SERVER_PORT
+
+  # We need this variable for the setenv
+  export DEPLOYMENT_ONLYOFFICE_HTTP_PORT
+  export DEPLOYMENT_CMIS_HTTP_PORT
+  do_start_exo_docker
+  )
+}
+
 #
 # Function that stops the app server
 #
 do_stop() {
+  if ${DEPLOY_EXO_DOCKER}; then
+   do_docker_stop
+  else
+   do_default_stop
+  fi
+}
+
+do_default_stop() {
   if [ ! -e "${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}" ]; then
     echo_warn "${PRODUCT_NAME} ${PRODUCT_VERSION} isn't deployed !"
     echo_warn "The product cannot be stopped"
@@ -1689,10 +1721,29 @@ do_stop() {
   fi
 }
 
+do_docker_stop(){
+  do_stop_exo_docker
+  do_stop_ldap
+  do_stop_onlyoffice
+  do_stop_cmis
+  do_stop_database
+  do_stop_es
+  do_stop_chat_server
+}
+
 #
 # Function that undeploys (delete) the app server
 #
+
 do_undeploy() {
+  if ${DEPLOY_EXO_DOCKER}; then
+   do_docker_undeploy
+  else
+   do_default_undeploy
+  fi
+}
+
+do_default_undeploy() {
   if [ ! -e "${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}" ]; then
     echo_warn "${PRODUCT_NAME} ${PRODUCT_VERSION} isn't deployed !"
     echo_warn "The product cannot be undeployed"
@@ -1740,6 +1791,42 @@ do_undeploy() {
     rm ${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}
     )
   fi
+}
+
+do_docker_undeploy() {
+if [ ! -e "${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}" ]; then
+  echo_warn "${PRODUCT_NAME} ${PRODUCT_VERSION} isn't deployed !"
+  echo_warn "The product cannot be undeployed"
+  exit 0
+else
+(
+  # The server is supposed to be already deployed.
+  # We load its settings from the configuration
+  do_load_deployment_descriptor
+  # Stop the server
+  do_stop
+  if ${DEPLOYMENT_DATABASE_ENABLED}; then
+    do_drop_database
+  fi
+  do_drop_onlyoffice_data
+  do_drop_ldap_data
+  do_drop_cmis_data
+  do_drop_chat
+  do_drop_es_data
+  echo_info "Undeploying server ${PRODUCT_DESCRIPTION} ${PRODUCT_VERSION} ..."
+  do_stop_exo_docker
+  # Close firewall ports
+  do_ufw_close_port ${DEPLOYMENT_RMI_REG_PORT} "JMX RMI REG" ${ADT_DEV_MODE}
+  do_ufw_close_port ${DEPLOYMENT_RMI_SRV_PORT} "JMX RMI SRV" ${ADT_DEV_MODE}
+  do_ufw_close_port ${DEPLOYMENT_CRASH_SSH_PORT} "CRaSH SSH" ${ADT_DEV_MODE}
+  if ${DEPLOYMENT_ONLYOFFICE_DOCUMENTSERVER_ENABLED} ; then
+      # close firewall port for Onlyoffice documentserver only if addon was deployed
+    do_ufw_close_port ${DEPLOYMENT_ONLYOFFICE_HTTP_PORT} "OnlyOffice Documentserver HTTP" ${ADT_DEV_MODE}
+  fi      
+  # Delete the deployment descriptor
+  rm ${ADT_CONF_DIR}/${INSTANCE_KEY}.${ACCEPTANCE_HOST}
+)
+fi
 }
 
 #
